+++ CHANGELOG.md
@@ -1,2 +1,9 @@
 # Changelog
 This changelog follows [the "Changelog specification" 1.0.0 *beta 2*](https://github.com/YaFou/Changelog-specification).
+
+## [1.0.0](https://github.com/Fluwork/Deprecation/releases/tag/v1.0.0)
+> 2020-07-03
+
+### Added
+- `trigger_deprecation` function.
+- `DeprecationAssert` class: helper to test if a deprecation was triggered for testing.